refactor: sniffer module
